Refactor ZenithView rendering and resource management
Unify and modernize rendering logic for ZenithView across WinUI and Uno (WPF) by introducing Frame/Present/EnsureResources methods and a FrameDispatcher. Move all D3D interop code into the WinUI partial class and refactor D3DTexture for explicit sync and present control. For Uno, encapsulate rendering resources in a new Surface class. Centralize resource (re)creation and cleanup, simplify the event loop, and improve code clarity and maintainability. Move WinUI Output config to ZenithViewOutputs.cs.
