--- sources/Views/Zenith.NET.Views.WinUI/D3D.cs
@@ -1,48 +1,0 @@
-﻿#if WINDOWS
-using System.Diagnostics;
-using Silk.NET.Core.Native;
-using Silk.NET.Direct3D11;
-using Silk.NET.DXGI;
-
-namespace Zenith.NET.Views.WinUI;
-
-internal static unsafe class D3D
-{
-    public static ComPtr<IDXGIFactory2> Factory;
-
-    public static ComPtr<ID3D11Device> Device;
-
-    public static ComPtr<ID3D11DeviceContext> DeviceContext;
-
-    static D3D()
-    {
-        DXGI = DXGI.GetApi(null);
-        D3D11 = D3D11.GetApi(null);
-
-        Success(DXGI.CreateDXGIFactory2(0, out Factory));
-
-        Success(D3D11.CreateDevice(default(ComPtr<IDXGIAdapter>),
-                                   D3DDriverType.Hardware,
-                                   0,
-                                   (uint)CreateDeviceFlag.BgraSupport,
-                                   null,
-                                   0,
-                                   D3D11.SdkVersion,
-                                   ref Device,
-                                   null,
-                                   ref DeviceContext));
-    }
-
-    public static DXGI DXGI { get; }
-
-    public static D3D11 D3D11 { get; }
-
-    public static void Success(int result)
-    {
-        if (result is not 0)
-        {
-            Debug.WriteLine($"Direct3D operation failed. HRESULT: 0x{result:X8}");
-        }
-    }
-}
-#endif
