+++ sources/Views/Zenith.NET.Views/ZenithViewOutputs.cs
@@ -2,6 +2,13 @@
 
 public static class ZenithViewOutputs
 {
+    public static Output WinUI { get; } = new()
+    {
+        ColorAttachments = [PixelFormat.B8G8R8A8UNorm],
+        DepthStencilAttachment = PixelFormat.D24UNormS8UInt,
+        SampleCount = SampleCount.Count1
+    };
+
     public static Output WPF { get; } = new()
     {
         ColorAttachments = [PixelFormat.B8G8R8A8UNorm],
